Add files via upload
This repository contains a decentralized voting application built with Solidity, a programming language for smart contracts on the Ethereum blockchain. The app utilizes blockchain technology to ensure secure , transparent, and tamper-proof voting processes. It allows users to cast their votes remotely while maintaining anonymity and preventing fraud. The app features a user-friendly interface, secure voter authentication, and real-time vote counting. It is a proof-of-concept project to demonstrate the potential of blockchain technology in voting systems.